Remove TS and add PropTypes
