+++ babel.config.js
@@ -1,5 +1,5 @@
 // Used by jest
 module.exports = {
-  presets: ['@babel/preset-env', '@babel/preset-react', '@babel/preset-typescript'],
+  presets: ['@babel/preset-env', '@babel/preset-react'],
   plugins: [['@babel/plugin-transform-runtime']],
 };
