+++ jest.e2e.js
@@ -1,7 +1,7 @@
 const DEBUG = process.env.DEBUG;
 
 module.exports = {
-  testRegex: 'src/.*\\.e2e\\.ts$',
+  testRegex: 'src/.*\\.e2e\\.js$',
   preset: 'jest-playwright-preset',
   testEnvironmentOptions: {
     'jest-playwright': {
