+++ package.json
@@ -16,12 +16,6 @@
         "@storybook/addon-links": "^6.1.20",
         "@storybook/react": "^6.1.20",
         "@testing-library/react": "^11.2.5",
-        "@types/jest": "^26.0.20",
-        "@types/react": "^17.0.2",
-        "@types/react-dom": "^17.0.1",
-        "@types/styled-components": "^5.1.7",
-        "@typescript-eslint/eslint-plugin": "^4.15.1",
-        "@typescript-eslint/parser": "^4.15.1",
         "axe-playwright": "^1.1.7",
         "babel-jest": "^26.6.3",
         "babel-loader": "^8.2.2",
@@ -38,21 +32,20 @@
         "prettier": "^2.2.1",
         "react": "^17.0.1",
         "react-dom": "^17.0.1",
-        "storybook-addon-designs": "^6.0.0-alpha.2",
-        "typescript": "^4.1.5"
+        "storybook-addon-designs": "^6.0.0-alpha.2"
     },
     "peerDependencies": {
         "react": "^17.0.1",
         "react-dom": "^17.0.1"
     },
     "dependencies": {
+        "prop-types": "^15.7.2",
         "styled-components": "^5.2.1"
     },
     "scripts": {
         "build": "tsc",
         "build-storybook": "build-storybook",
-        "type:check": "tsc --noEmit",
-        "lint": "eslint . --ext .js,.jsx,.ts,.tsx --fix",
+        "lint": "eslint . --ext .js,.jsx --fix",
         "storybook": "start-storybook -p 6006",
         "test": "jest -c jest.config.js",
         "test:e2e": "cross-env DEBUG=true jest -c jest.e2e.js",
@@ -61,11 +54,11 @@
         "specify:pull": "specify pull"
     },
     "husky": {
-      "hooks": {
-        "pre-commit": "lint-staged"
-      }
+        "hooks": {
+            "pre-commit": "lint-staged"
+        }
     },
     "lint-staged": {
-      "*.{js,jsx,ts,tsx}": "eslint --ext .js,.jsx,.ts,.tsx --fix"
+        "*.{js,jsx}": "eslint --ext .js,.jsx --fix"
     }
 }
