+++ src/Button/Button.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const Button = styled.button`
+  background: ${(props) => props.theme.color.avocado100};
+`;
