+++ src/Button/Button.stories.js
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { Meta } from '@storybook/react/types-6-0';
 import { withDesign } from 'storybook-addon-designs';
 import { Button } from './Button';
 
@@ -18,4 +17,4 @@ export default {
   title: 'Button',
   component: Button,
   decorators: [withDesign],
-} as Meta;
+};
