+++ src/ThemeProvider/ThemeProvider.js
@@ -1,0 +1,12 @@
+import * as React from 'react';
+import PropTypes from 'prop-types';
+import { ThemeProvider as StyledThemeProvider } from 'styled-components';
+
+export const ThemeProvider = ({ children, theme }) => {
+  return <StyledThemeProvider theme={theme}>{children}</StyledThemeProvider>;
+};
+
+ThemeProvider.propTypes = {
+  children: PropTypes.node.isRequired,
+  theme: PropTypes.object.isRequired,
+};
