+++ src/ThemeProvider/hooks.js
@@ -1,0 +1,5 @@
+import { useTheme as useStyledTheme } from 'styled-components';
+
+export const useTheme = () => {
+  return useStyledTheme();
+};
