--- src/ThemeProvider/hooks.ts
@@ -1,6 +1,0 @@
-import { useTheme as useStyledTheme } from 'styled-components';
-import { Theme } from './types';
-
-export const useTheme = (): Theme => {
-  return useStyledTheme() as Theme;
-};
