+++ src/ThemeProvider/index.js
@@ -1,3 +1,2 @@
 export * from './ThemeProvider';
-export * from './types';
 export * from './hooks';
