--- src/ThemeProvider/types.ts
@@ -1,8 +1,0 @@
-export interface Theme {
-  // Specify doesn't generate types for the moment
-  color: { [key: string]: string };
-}
-
-export interface ThemeProps {
-  theme: Theme;
-}
