--- src/global.d.ts
@@ -1,7 +1,0 @@
-import { Browser, Page } from 'playwright';
-
-declare global {
-  const page: Page;
-  const browser: Browser;
-  const browserName: string;
-}
