--- src/themes/light-theme.ts
@@ -1,45 +1,0 @@
-export const lightTheme = {
-    color: {
-        aqua100: '#e5f3ff',
-        aqua200: '#b8e1ff',
-        aqua500: '#66b7f1',
-        aqua600: '#009ee0',
-        aqua700: '#2684b8',
-        avocado100: '#eafbe7',
-        avocado200: '#c6f0c2',
-        avocado500: '#5cb176',
-        avocado600: '#338648',
-        avocado700: '#2f6846',
-        carbon0: '#ffffff',
-        carbon100: '#f6f6f9',
-        carbon150: '#eaeaef',
-        carbon200: '#dcdce4',
-        carbon300: '#c0c0cf',
-        carbon400: '#a5a5ba',
-        carbon500: '#8e8ea9',
-        carbon600: '#666687',
-        carbon700: '#4a4a6a',
-        carbon800: '#32324d',
-        carbon900: '#212134',
-        cherry100: '#fcecea',
-        cherry200: '#f5c0b8',
-        cherry500: '#ee5e52',
-        cherry600: '#dd2b23',
-        cherry700: '#b72b1a',
-        elixir100: '#f0f0ff',
-        elixir200: '#d9d8ff',
-        elixir500: '#7b79ff',
-        elixir600: '#4945ff',
-        elixir700: '#271fe0',
-        grape100: '#f6ecfc',
-        grape200: '#e0c1f4',
-        grape500: '#ac73e6',
-        grape600: '#9d41f1',
-        grape700: '#8724cb',
-        pineapple100: '#fdf4dc',
-        pineapple200: '#fae7b9',
-        pineapple500: '#fab861',
-        pineapple600: '#f29d41',
-        pineapple700: '#d9822f',
-    },
-};
